+++ src/process.h
@@ -2,13 +2,14 @@
 
 #include <stdio.h>
 #include <unistd.h>
+#include <stdint.h>
 
 typedef struct {
     int color;
     pid_t pid;
     int fd;
     FILE* f;
-    int out_fd_count;
+    uint8_t out_fd_count;
     int *out_fds;
     char *name;
     char *cmd;
@@ -26,10 +27,10 @@ void process_stop(Process *p);
 
 void process_add_output_fd(Process *p, int fd);
 
-void process_remove_output_fd_at(Process *p, int index);
+void process_remove_output_fd_at(Process *p, size_t index);
 
 int process_print_status(const Process* p);
 
-int process_serialize(Process *p, char* buffer, int buf_size);
+size_t process_serialize(Process *p, char* buffer, size_t buf_size);
 
-int process_deserialize(char* buffer, Process* p);
+size_t process_deserialize(char* buffer, Process* p);
